refactor: moved pages to test cases folder
