+++ ui/src/app/app-routing.module.ts
@@ -16,9 +16,9 @@ import { LoginComponent } from './pages/login/login.component';
 import { SettingsComponent } from './components/settings/settings.component';
 import { AuthLayoutComponent } from './layouts/auth-layout/auth-layout.component';
 import { MainLayoutComponent } from './layouts/main-layout/main-layout.component';
-import { TestCasesComponent } from './pages/test-cases/test-cases.component';
-import { TestCaseDetailPageComponent } from './pages/test-case-detail/test-case-detail.component';
-import { TestCasesHomeComponent } from './pages/test-cases-home/test-cases-home.component';
+import { TestCaseListComponent } from './pages/test-cases/test-case-list/test-case-list.component';
+import { TestCaseDetailPageComponent } from './pages/test-cases/test-case-detail/test-case-detail.component';
+import { TestCaseHomeComponent } from './pages/test-cases/test-case-home/test-case-home.component';
 
 const routes: Routes = [
   {
@@ -147,7 +147,7 @@ const routes: Routes = [
       },
       {
         path: 'test-cases/:userStoryId',
-        component: TestCasesComponent,
+        component: TestCaseListComponent,
         canActivate: [UserGuard],
       },
       {
@@ -172,7 +172,7 @@ const routes: Routes = [
       },
       {
         path: 'test-cases-home',
-        component: TestCasesHomeComponent,
+        component: TestCaseHomeComponent,
         canActivate: [UserGuard],
         data: {
           breadcrumb: {
