+++ ui/src/app/pages/test-cases/test-case-detail/test-case-detail.component.scss
