--- ui/src/app/pages/test-cases/test-cases.component.scss
@@ -1,1 +1,0 @@
-// Add any specific styles for the test cases component here
